feat: add a "configuration issues" button + dialog to the global nav bar in development
