+++ packages/sanity/src/studio/screens/schemaErrors/utils.ts
@@ -1,0 +1,18 @@
+import {SchemaValidationProblemPath} from '@sanity/types'
+
+export function renderPath(path: SchemaValidationProblemPath) {
+  return path
+    .map((segment) => {
+      if (segment.kind === 'type') {
+        return `${segment.name || '<unnamed>'}(${segment.type})`
+      }
+
+      if (segment.kind === 'property') {
+        return segment.name
+      }
+
+      return null
+    })
+    .filter(Boolean)
+    .join(' > ')
+}
